chore: reorganise types
