+++ src/main/@types/aurora-tools.ts
@@ -1,0 +1,14 @@
+import { AwsConfig } from './common';
+
+/**
+ * Configuration options for the Aurora toolkit
+ * @export
+ * @interface AuroraConfig
+ * @extends {AwsConfig}
+ */
+export interface AuroraConfig extends AwsConfig {
+  clusterNameA: string;
+  clusterNameB: string;
+  minimumClusterSize: number;
+  skipDeleteSnapshots: boolean;
+}
