+++ src/main/@types/common.ts
