+++ src/main/@types/dynamo-tools.ts
@@ -1,0 +1,13 @@
+import { AwsConfig } from './common';
+
+/**
+ * Configuration options for the Aurora toolkit
+ * @export
+ * @interface DynamoConfig
+ * @extends {AwsConfig}
+ */
+export interface DynamoConfig extends AwsConfig {
+  tableNameA: string;
+  tableNameB: string;
+  waitForTableDelete?: boolean;
+}
